--- src/main/java/com/example/ottokeng/domain/lost/entity/Get.java
@@ -1,6 +1,0 @@
-package com.example.ottokeng.domain.lost.entity;
-
-public enum Get {
-    GET,
-    NOT_GET
-}
