--- src/main/java/com/example/ottokeng/domain/lost/utils/PostUtil.java
@@ -1,4 +1,0 @@
-package com.example.ottokeng.domain.lost.utils;
-
-public class PostUtil {
-}
