+++ src/main/java/com/example/ottokeng/domain/post/entity/Get.java
@@ -1,0 +1,6 @@
+package com.example.ottokeng.domain.post.entity;
+
+public enum Get {
+    GET,
+    NOT_GET
+}
