+++ src/main/java/com/example/ottokeng/domain/post/presentation/PostController.java
@@ -1,9 +1,9 @@
-package com.example.ottokeng.domain.lost.presentation;
+package com.example.ottokeng.domain.post.presentation;
 
-import com.example.ottokeng.domain.lost.presentation.dto.request.PostWritingRequest;
-import com.example.ottokeng.domain.lost.presentation.dto.request.ModifyPostWritingRequest;
-import com.example.ottokeng.domain.lost.presentation.dto.response.AllPostsResponse;
-import com.example.ottokeng.domain.lost.service.PostService;
+import com.example.ottokeng.domain.post.presentation.dto.request.PostWritingRequest;
+import com.example.ottokeng.domain.post.presentation.dto.request.ModifyPostWritingRequest;
+import com.example.ottokeng.domain.post.presentation.dto.response.AllPostsResponse;
+import com.example.ottokeng.domain.post.service.PostService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,30 +12,30 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/post")
-public class LostController {
-    private final PostService lostWritingService;
+public class PostController {
+    private final PostService postService;
 
     @GetMapping("/writing")
     public ResponseEntity<AllPostsResponse> showlost(){
-        AllPostsResponse showLostsResponse = lostWritingService.getAllPost();
-        return new ResponseEntity<>(showLostsResponse, HttpStatus.OK);
+        AllPostsResponse allPost = postService.getAllPost();
+        return new ResponseEntity<>(allPost, HttpStatus.OK);
     }
 
     @PostMapping("/writing")
     public ResponseEntity<Void> postWriting(@RequestBody PostWritingRequest request){
-        lostWritingService.postWritingExecute(request);
+        postService.postWritingExecute(request);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     @PatchMapping("/writing/{id}")
     public ResponseEntity<Void> patchWriting(@PathVariable Long id, @RequestBody ModifyPostWritingRequest request){
-        lostWritingService.patchWritingExecutte(id, request);
+        postService.patchWritingExecutte(id, request);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
     @DeleteMapping("/writing")
     public ResponseEntity<Void> deleteWriting(@RequestParam Long id){
-        lostWritingService.deleteWritingExecute(id);
+        postService.deleteWritingExecute(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
