+++ src/main/java/com/example/ottokeng/domain/post/presentation/dto/request/PostWritingRequest.java
@@ -1,7 +1,7 @@
-package com.example.ottokeng.domain.lost.presentation.dto.request;
+package com.example.ottokeng.domain.post.presentation.dto.request;
 
-import com.example.ottokeng.domain.lost.entity.Get;
-import com.example.ottokeng.domain.lost.entity.Type;
+import com.example.ottokeng.domain.post.entity.Get;
+import com.example.ottokeng.domain.post.entity.Type;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
