+++ src/main/java/com/example/ottokeng/domain/post/repository/PostRepository.java
@@ -1,6 +1,6 @@
-package com.example.ottokeng.domain.lost.repository;
+package com.example.ottokeng.domain.post.repository;
 
-import com.example.ottokeng.domain.lost.entity.Post;
+import com.example.ottokeng.domain.post.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
