+++ src/main/java/com/example/ottokeng/domain/post/utils/PostUtil.java
@@ -1,0 +1,4 @@
+package com.example.ottokeng.domain.post.utils;
+
+public class PostUtil {
+}
