+++ src/main/java/com/example/ottokeng/domain/user/entity/User.java
@@ -1,7 +1,7 @@
 package com.example.ottokeng.domain.user.entity;
 
 
-import com.example.ottokeng.domain.lost.entity.Post;
+import com.example.ottokeng.domain.post.entity.Post;
 import com.example.ottokeng.global.enumType.Role;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
