+++ src/main/java/com/example/ottokeng/global/exception/ErrorCode.java
@@ -13,8 +13,7 @@ public enum ErrorCode {
     USER_NOT_FOUND(404, "존재하지 않는 유저입니다."),
     UNABLE_TO_ISSUANCE_REFRESHTOKEN(400, "유효하지 않은 리프레쉬 토큰입니다."),
     ALREADY_BLACKLIST(409, "이미 블랙리스트에 존재합니다."),
-    FIND_WRITING_NOT_FOUND(404, "존재하지 않는 찾기 글입니다."),
-    LOST_WRITING_NOT_FOUND(404, "존재하지 않는 등록 글입니다.");;
+    POST_NOT_FOUND(404, "존재하지 않는 게시글입니다.");;
 
     private final int status;
     private final String detail;
